custom_container/requirements.txt
numpy
scikit-learn
google-cloud-storage
uvicorn[standard]
gunicorn
fastapi[all]
